// src/types/changeUserDataRequest.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * An explanation about the purpose of this instance.
 */
export type TheFirstAnyOfSchema = string;
/**
 * An explanation about the purpose of this instance.
 */
export type TagsSchema = TheFirstAnyOfSchema[];
/**
 * An explanation about the purpose of this instance.
 */
export type ShotTypesSchema = ("Landing" | "Web Application" | "Website" | "Mobile")[];

/**
 * The root schema comprises the entire JSON document.
 */
export interface ChangeUserDataRequest {
  tags: TagsSchema;
  shotTypes: ShotTypesSchema;
}